+++ libraries/MCP4261/MCP4261.cpp
@@ -1,7 +1,7 @@
 //
 //    FILE: MCP4261.cpp
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.1.0
+// VERSION: 0.2.0
 //    DATE: 2024-02-21
 // PURPOSE: Arduino library for MCP4261 SPI based digital potentiometers.
 //     URL: https://github.com/RobTillaart/MCP4261
@@ -10,18 +10,21 @@
 #include "MCP4261.h"
 
 
-//  see page 18 datasheet
-#define MCP4261_IGNORE_CMD        0x00
-#define MCP4261_WRITE_CMD         0x00
-#define MCP4261_SHUTDOWN_CMD      0x20
-#define MCP4261_NONE_CMD          0x30
+#define MCP4261_CMD_WRITE         0x00
+#define MCP4261_CMD_INCR          0x04
+#define MCP4261_CMD_DECR          0x08
+#define MCP4261_CMD_READ          0x0C
+
+#define MCP4261_REG_TCON          0x04
+#define MCP4261_REG_STATUS        0x05
+#define MCP4261_REG_EEPROM(x)     (0x06 + x)
 
 
   //       HARDWARE SPI
 MCP4261::MCP4261(uint8_t select, uint8_t shutdown, __SPI_CLASS__ * mySPI)
 {
   _pmCount  = 2;
-  _maxValue = 255;
+  _maxValue = MCP42XX_MAX_VALUE;
   _select   = select;
   _shutdown = shutdown;
   _dataIn   = 255;
@@ -37,7 +40,7 @@ MCP4261::MCP4261(uint8_t select, uint8_t shutdown, __SPI_CLASS__ * mySPI)
 MCP4261::MCP4261(uint8_t select, uint8_t shutdown, uint8_t dataIn, uint8_t dataOut, uint8_t clock)
 {
   _pmCount  = 2;
-  _maxValue = 255;
+  _maxValue = MCP42XX_MAX_VALUE;
   _select   = select;
   _shutdown = shutdown;
   _dataIn   = dataIn;
@@ -48,7 +51,7 @@ MCP4261::MCP4261(uint8_t select, uint8_t shutdown, uint8_t dataIn, uint8_t dataO
 }
 
 
-void MCP4261::begin(uint16_t value)
+void MCP4261::begin()
 {
   pinMode(_select, OUTPUT);
   digitalWrite(_select, HIGH);
@@ -71,7 +74,9 @@ void MCP4261::begin(uint16_t value)
     digitalWrite(_dataOut, LOW);
     digitalWrite(_clock,   LOW);
   }
-  reset(value);
+  //  get initial values from device.
+  getValueDevice(0);
+  getValueDevice(1);
 }
 
 
@@ -82,6 +87,11 @@ void MCP4261::reset(uint16_t value)
   setValue(value);  //  set all to same value.
 }
 
+
+uint8_t MCP4261::pmCount()
+{
+  return _pmCount;
+}
 
 
 /////////////////////////////////////////////////////////////////////////////
@@ -102,10 +112,9 @@ bool MCP4261::setValue(uint8_t pm, uint16_t value)
   if (value > _maxValue) return false;
   _value[pm] = value;
 
-  uint8_t cmd = 0x00;
-  if (pm == 1) cmd = 0x10;
-  if (value > 0xFF) cmd |= (value >> 8);  //  high bits
-  writeDevice(2, cmd, value & 0xFF);
+  uint8_t value1 = MCP4261_CMD_WRITE | (pm << 4);
+  if (value > 0xFF) value1 |= (value >> 8);  //  high bits
+  writeRegister2(value1, value & 0xFF);
   return true;
 }
 
@@ -117,15 +126,23 @@ uint16_t MCP4261::getValue(uint8_t pm)
 }
 
 
+uint16_t MCP4261::getValueDevice(uint8_t pm)
+{
+  if (pm >= _pmCount) return 0;
+  uint8_t value1 = MCP4261_CMD_READ | (pm << 4);
+  _value[pm] = readRegister(value1);
+  return _value[pm];
+}
+
+
 bool MCP4261::incrValue(uint8_t pm)
 {
   if (pm >= _pmCount) return false;
   if (_value[pm] >= _maxValue) return false;
   _value[pm]++;
 
-  uint8_t cmd = 0x04;
-  if (pm == 1) cmd = 0x14;
-  writeDevice(1, cmd, cmd);  //  value2 = DUMMY
+  uint8_t value1 = MCP4261_CMD_INCR | (pm << 4);
+  writeRegister1(value1);
   return true;
 }
 
@@ -136,9 +153,8 @@ bool MCP4261::decrValue(uint8_t pm)
   if (_value[pm] == 0) return false;
   _value[pm]--;
 
-  uint8_t cmd = 0x08;
-  if (pm == 1) cmd = 0x18;
-  writeDevice(1, cmd, cmd);  //  value2 = DUMMY
+  uint8_t value1 = MCP4261_CMD_DECR | (pm << 4);
+  writeRegister1(value1);
   return true;
 }
 
@@ -153,24 +169,81 @@ bool MCP4261::setValueNV(uint8_t pm, uint16_t value)
   if (pm >= _pmCount) return false;
   if (value > _maxValue) return false;
 
-  uint8_t cmd = 0x20;
-  if (pm == 1) cmd = 0x30;
-  if (value > 0xFF) cmd |= (value >> 8);  //  high bits
-  writeDevice(2, cmd, value & 0xFF);
+  uint8_t value1 = MCP4261_CMD_WRITE | ((pm + 2) << 4);
+  if (value > 0xFF) value1 |= (value >> 8);  //  high bits
+  writeRegister2(value1, value & 0xFF);
   return true;
+}
+
+
+uint16_t MCP4261::getValueNV(uint8_t pm)
+{
+  if (pm >= _pmCount) return 0;
+  uint8_t value1 = MCP4261_CMD_READ | ((pm + 2) << 4);
+  return readRegister(value1);
 }
 
 
 /////////////////////////////////////////////////////////////////////////////
 //
-//  MISC
+//  TERMINAL CONTROL
 //
-uint8_t MCP4261::pmCount()
+void MCP4261::setTCONMask(uint16_t mask)
 {
-  return _pmCount;
+  uint8_t value1 = MCP4261_CMD_WRITE | (MCP4261_REG_TCON << 4);
+  mask &= 0xFF;  //  only lower 8 bits used.
+  writeRegister2(value1, mask);
 }
 
 
+uint16_t MCP4261::getTCONMask()
+{
+  uint8_t value1 = MCP4261_CMD_READ | (MCP4261_REG_TCON << 4);
+  uint16_t mask = readRegister(value1);
+  return mask & 0x01FF;  //  0x1FF to get bit 8 which is always HIGH.
+}
+
+
+/////////////////////////////////////////////////////////////////////////////
+//
+//  STATUS
+//
+uint16_t MCP4261::getStatusMask()
+{
+  uint8_t value1 = MCP4261_CMD_READ | (MCP4261_REG_STATUS << 4);
+  uint16_t mask = readRegister(value1);
+  return mask & 0x1F;
+}
+
+
+/////////////////////////////////////////////////////////////////////////////
+//
+//  EEPROM
+//
+bool MCP4261::setEEPROM(uint8_t index, uint16_t value)
+{
+  if (index > 9) return false;
+  if (value > 0x1FF) return false;  //  511
+  uint8_t value1 = MCP4261_CMD_WRITE | (MCP4261_REG_EEPROM(index) << 4);
+  if (value > 0xFF) value1 |= (value >> 8);
+  writeRegister2(value1, value & 0xFF);
+  return true;
+}
+
+
+uint16_t MCP4261::getEEPROM(uint8_t index)
+{
+  if (index > 9) return 0;  //  0xFFFF ?
+  uint8_t value1 = MCP4261_CMD_READ | (MCP4261_REG_EEPROM(index) << 4);
+  uint16_t value = readRegister(value1);
+  return value;
+}
+
+
+/////////////////////////////////////////////////////////////////////////////
+//
+//  POWER
+//
 void MCP4261::powerOn()
 {
   digitalWrite(_shutdown, HIGH);
@@ -217,48 +290,86 @@ bool MCP4261::usesHWSPI()
 //  PROTECTED
 //
 
-/*
-TODO - see page 46
-- need read8, read16
-- bid8, bidi16
-*/
-
-
 //
 //  USES SPI MODE 0
 //
-void MCP4261::writeDevice(uint8_t count, uint8_t value1, uint8_t value2)
+void MCP4261::writeRegister1(uint8_t value1)
 {
   digitalWrite(_select, LOW);
   if (_hwSPI)
   {
     _mySPI->beginTransaction(_spi_settings);
     _mySPI->transfer(value1);
-    if (count == 2) _mySPI->transfer(value2);
     _mySPI->endTransaction();
   }
   else      //  Software SPI
   {
     swSPI_write(value1);
-    if (count == 2) swSPI_write(value2);
   }
   digitalWrite(_select, HIGH);
 }
 
 
+void MCP4261::writeRegister2(uint8_t value1, uint8_t value2)
+{
+  digitalWrite(_select, LOW);
+  if (_hwSPI)
+  {
+    _mySPI->beginTransaction(_spi_settings);
+    _mySPI->transfer(value1);
+    _mySPI->transfer(value2);
+    _mySPI->endTransaction();
+  }
+  else      //  Software SPI
+  {
+    swSPI_write(value1);
+    swSPI_write(value2);
+  }
+  digitalWrite(_select, HIGH);
+}
+
+
+uint16_t MCP4261::readRegister(uint8_t value)
+{
+  uint16_t rv = 0;
+  digitalWrite(_select, LOW);
+  if (_hwSPI)
+  {
+    _mySPI->beginTransaction(_spi_settings);
+    rv += _mySPI->transfer(value);
+    rv <<= 8;
+    rv += _mySPI->transfer(0x00);
+    _mySPI->endTransaction();
+  }
+  else      //  Software SPI
+  {
+    rv += swSPI_write(value);
+    rv <<= 8;
+    rv += swSPI_write(0x00);
+  }
+  digitalWrite(_select, HIGH);
+  rv &= 0x01FF;  //  9 bits data only
+  return rv;
+}
+
+
 //  MSBFIRST
-void MCP4261::swSPI_write(uint8_t value)
+uint8_t MCP4261::swSPI_write(uint8_t value)
 {
   uint8_t clk = _clock;
   uint8_t dao = _dataOut;
-  
+  uint8_t dai = _dataIn;
+  uint8_t val = 0;
+
   //  MSBFIRST
   for (uint8_t mask = 0x80; mask; mask >>= 1)
   {
     digitalWrite(dao,(value & mask));
     digitalWrite(clk, HIGH);
+    if (digitalRead(dai)) val |= mask;
     digitalWrite(clk, LOW);
   }
+  return val;
 }